requirements.txt
numpy==2.2.5
scikit-learn==1.6.1
pandas==2.2.3
joblib==1.5.0
apache-airflow==3.0.0
dotenv==0.9.9
requests==2.32.3
mlflow==2.22.0
clearml==1.18.0
wandb==0.19.11
dvc==3.42.0
dvc-s3==3.0.1
fastapi==0.115.11
xgboost==3.0.2

